% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{batchGenesFile}
\alias{batchGenesFile}
\title{Get batch gene information from a file.}
\usage{
batchGenesFile(
  fileName,
  species,
  release,
  header = FALSE,
  details = FALSE,
  debug = FALSE
)
}
\arguments{
\item{fileName}{A file name with ensembl ids}

\item{species}{Either "Mm" or "Hs"}

\item{release}{ensimpl release, see releases()}

\item{header}{TRUE if the file has a header line.}

\item{details}{TRUE for extra information.}

\item{debug}{Display some debugging information.}
}
\value{
By default, with details=FALSE, a tibble will be returned.
  If details=TRUE, a list will be returned with each element being
  the matching "id".
}
\description{
Get batch gene information from a file.
}
