% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{externalIDs}
\alias{externalIDs}
\title{Lookup by ids.}
\usage{
externalIDs(ids, species, release, source_db = "Ensembl", debug = FALSE)
}
\arguments{
\item{ids}{A list of Ensembl identifiers.}

\item{species}{Either "Mm" or "Hs"}

\item{release}{ensimpl release, see releases()}

\item{source_db}{A string which defaults to 'Ensembl', but could also
be one of the source_db values from externalDBs() function.}

\item{debug}{Display some debugging information.}
}
\value{
A tibble with the external ids listed.
}
\description{
Lookup by ids.
}
\examples{
extids <- externalIDs(c("ENSMUSG00000000001", "ENSMUSG00000000002"),
                      "Mm", "101")
}
