% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{getGene}
\alias{getGene}
\title{Get gene information.}
\usage{
getGene(id, species, release, details = FALSE, debug = FALSE)
}
\arguments{
\item{id}{The Ensembl identifier.}

\item{species}{Either "Mm" or "Hs"}

\item{release}{ensimpl release, see releases()}

\item{details}{TRUE for extra information including transcripts.}

\item{debug}{Display some debugging information.}
}
\value{
A list with gene information.
}
\description{
Get gene information.
}
\examples{
gene <- getGene("ENSMUSG00000000001", species="Mm", release="101")
}
