% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{getGeneHistory}
\alias{getGeneHistory}
\title{Get gene history information.}
\usage{
getGeneHistory(id, species, release_start, release_end, debug = FALSE)
}
\arguments{
\item{id}{The Ensembl identifier.}

\item{species}{Either "Mm" or "Hs"}

\item{release_start}{ensimpl release, see releases()}

\item{release_end}{ensimpl release, see releases()}

\item{debug}{Display some debugging information.}

\item{details}{TRUE for extra information including transcripts.}
}
\value{
A list with gene information.
}
\description{
Get gene history information.
}
\examples{
gene <- getGeneHistory("ENSMUSG00000000001", species="Mm",
                       release_start="50", release_end="100")
}
