% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{searchGenes}
\alias{searchGenes}
\title{Search for gene information.}
\usage{
searchGenes(term, species, release, debug = FALSE)
}
\arguments{
\item{term}{The search term.}

\item{species}{Either "Mm" or "Hs"}

\item{release}{ensimpl release, see releases()}

\item{debug}{Display some debugging information.}
}
\description{
Search for gene information.
}
\examples{
results <- searchGenes("Ace", "Mm", "104")
}
